Bug#19792203 - FAILING ASSERTION: TRX->IS_DD_TRX == FALSE WHEN KILLING QUERY WITH CONVERT_TZ
When the SQL-function CONVERT_TZ is executed the function my_tz_find
is called to get timezone information. This function calls
the function open_trans_system_tables_for_read() to open the tables
time_zone_name, time_zone, time_zone_transition_type, time_zone_transition.
This function starts attachable data dictionary transaction and calls the
function open_tables() to iterate along the list of tables mentioned above
and call the method open() for every table's handlerton.
Handlerton for InnoDB tables sets the flag trx->is_dd_trx on opening the table.
After all tables are opened SUCCESSFULLY the function lock_tables() is called
in order to lock the tables. As a side effect this function sets the THD::lock
to point to a lock structure. Later on closing system tables the value of
THD::lock is checked to determine whether the table's lock has to be released.
When THD::lock is not NULL the method ha_innodbase::external_lock() is called
to release the InnoDB table lock. Implementaion of ha_innodbase::external_lock()
makes transaction commit and resets the flag trx->is_dd_trx.
The reason for the bug is that if an attempt to open the second or further
table is failed the function lock_tables() is never called and the transaction
never be closed. As side effect the flag trx->is_dd_trx is never be reset that
leads to crash inside the function trx_free().

To fix the crash setting of the flag trx->is_dd_trx moved into the method
ha_innobase::external_lock().