Bug#18661573 : MIN VALUE FOR STORED_PROGRAM_CACHE TOO HIGH,                DEFAULT MAX PACKET = HUGE SP MEMORY
The core issue here is, when stored programs are big then there
is chance of stored program caches occupying lot of memory.
Even with the minimum stored_program_cache value there is a
chance of facing this situation.

The problem in addressing the core issue is that, currently there
is no way to limit the memory consumption for the stored programs.
We have system variable "stored_program_cache", which limits the
number of stored programs, but not the memory used for them.
So, this core issue will be addressed by a separate WL.

As part of this patch, after discussing with support, changing
the min value of system variable "stored_program_cache" to 16
from 256. Reducing number of SP in caches will help in reducing
total amount of memory consumed by SP caches when there is
chance of having SPs consuming large memory.

With this change "stored_program_cache" system variables min/
default/max value will be as below,
        min value     : 16
        default value : 256
        max value     : 512 * 1024