Bug#20266847 : PURGE THREAD SHOULD CHECK PURGE_SYS->STATE AFTER EVERY BATCH
purge_sys->state is set to PURGE_STATE_STOP to stop the purge when exporting
a tablespace. This flag is checked by the the purge thread in
srv_purge_coordinator_suspend(). The purge thread goes to this function after
finishing a full sweep of purge in srv_do_purge(). If we have a large history
list, a full purge can take quite a while.

Approved by Sunny in IM