Initialize ut_stage_alter_t::m_progress to NULL
WL#5889 Add InnoDB events to Performance Schema's Event Stage table

It is possible that an ut_stage_alter_t object is constructed,
begin_phase_read_pk() method is never called and then the
begin_phase_end() and the destructor are called. This can happen if OOM
occurs in ha_innobase::inplace_alter_table(). Given that
begin_phase_read_pk() initializes the m_progress member, in this
scenario it will remain uninitialized for the calls to begin_phase_end()
and the destructor (which can properly handle it if it is NULL).

Thus initialize m_progress to NULL in the constructor.