Description: According to the manual, the maximum value for group_concat_max_len in 64bit mysqld is 18446744073709547520. This works fine when set dynamically in the mysql command prompt. However, if the value is set in my.cnf, mysqld will refuse to honor and will not start up.
Analysis: When the option group_concat_max_len is parsed and tries to store the given value,
it will call eval_num_suffix() to convert the given value string to numeric value with the
string function strtoll(). Since the give value won't fit in the longlong type an error is
thrown and server refuse to start.

Fix: As a fix we need to use strtoull() instead of strtoll(). which can accumulate the given
unsigned longlong value.

Note: This bug is already fixed in mysql-5.7 under bug #16491145. Hence used the same code changes to fix this bug as well.