Bug #20316320 - DEBUG BUILD ON WINDOWS SHOULD REPLACE /OB0 WITH /OB1 TO SPEED UP TEST RUNS
Replace /Ob0 which CMake automatically adds for debug builds, with /Ob1

This caused linking error due to some functions in my_time.c which
were declared inline there but not in my_time.h. This makes little
sense, so I removed the inline.

Added new CMake option WIN_DEBUG_NO_INLINE which is default OFF.
If this is turned ON, inlining will again be disabled as before.