Bug#20125466 - ASSERT IN FIL_IO BUF_READ_PAGE_LOW ROW_UPDATE_FOR_MYSQL
After a cascade update, we will update the statstics for referenced tables.
And we don't acquire any lock as we want better performance. So there
could be some 'DROP TABLE' thread which would marks all the root page numbers
of the referenced table's indexes as FIL_NULL. So the updater thread
could read an root page with number of FIL_NULL and read the page out of range.

We don't have MDL locks for referenced tables in the server layer now,
so we only protect the conflicts of cascade update and 'DROP TABLE' by
our dict_operation_lock. But this is released before the statistics updated.
Considering the performance issue, we could skip the lock but do a
double check when the statistics updater accesses the index.

Reviewed-by: Jimmy <jimmy.yang@oracle.com>
RB: 7735