WL#8190: Refactor low-level thread handling
This patch refactors the low-level code for handling OS threads.
It unifies the API for creating and joining threads so that there
no longer are separate functions for Windows and non-Windows
platforms.

This also makes sure that thread creation on Windows is always
tracked properly by performance schema.

The patch replaces macros with static inline functions or typedefs
where appropriate and does general refactoring of thread handling
and mysys initialization code.