Bug#20302542 FIX #INCLUDE DEPENDENCIES FOR PLUGIN.H AND PLUGIN_FTPARSER.H
The file include/mysql/plugin.h currently does
  include "plugin_ftparser.h"
This means that the ABI check for all plugins
depend on the fulltext plugin ABI.
This doesn't scale and is hard to maintain if
we want to add more plugin interfaces.

Fix:
Remove that #include.
Files which depend on plugin_ftparser.h should #include it.