Bug#20330831 INCONSISTENT ERROR REPORTING ABOUT PERSISTENT STATISTICS TABLES
The file dict0stats.cc defines two sets of symbols, for
internal and external names of the InnoDB tables
that store persistent statistics.

The names with _PRINT suffix are indented for use in diagnostic output.
However, there are two messages that use the internal names
(the symbols with no _PRINT suffix).

Reviewed-by: Vasil Dimov <vasil.dimov@oracle.com>