Improve garbage collection of clusterj artifacts
A user object obtained via query or session.newInstance is a proxy that
holds a strong reference to an NdbRecordSmartValueHandlerImpl, which 
in turn holds a strong reference to an NdbRecordOperationImpl, which 
holds strong references to ByteBuffers which wrap native byte arrays. 

When the user object goes out of scope, it is garbage collected but 
the objects that are now unreferenced take a long time to be collected.

This patch immediately releases objects that are referenced by user objects 
when the user object is collected. This can improve the ability of the 
garbage collector to release the space used by the objects.

DynamicObject.java:
  implement finalize method for dynamically created objects

DynamicObjectDelegate.java:
  add finalize to the methods to be implemented by the dynamic object

DomainTypeHandlerImpl.java:
  require that the proxy (user object) implement finalize

InvocationHandlerImpl.java:
  implement release

KeyValueHandlerImpl.java:
  implement release

NdbRecordBlobImpl.java:
  implement release() to release unneeded objects

NdbRecordOperationImpl.java:
  implement release() to release unneeded objects

NdbRecordSmartValueHandlerImpl.java:
  implement finalize() which forwards to release() to release unneeded objects

NdbOpenJPAValueHandler.java:
  implement release

NoPublicConstructorClusterConnectionService.java:
  fix bug in test