WL#6833: GCS Replication: Read-set free Certification Module (DBSM Snapshot Isolation)
Add GTID_NEXT support to Group Replication certifier.

Certification algorithm now compares full transaction snapshot
version instead of comparing the last GNO without intervals.
This means that when a transaction snapshot version (Gtid_set) is a
subset of a parallel transaction snapshot version a conflict is
detected and the last one to be certified is cancelled.