Fixed bug#20328786: WL411:VALGRIND WARNINGS OF CONDITIONAL JUMP WHILE SELECTING FROM VIEW
There is an inconsistency between readset of derived table and selected
readset of view. This is because of missing to treat stored GC as normal
columns in function insert_fields during setup_wild. However such a treatment
is done in opening a derived table. Such a consistency results in this bug.