Bug #19306793 	CORE IN NDBAPI WHEN EXTENDED EXTENSION TABLE IS CREATED IN MYSQL 7.4
The increase in Max hashmap buckets from 240 to 3480 increased the 
size of the DictHashMapInfo::HashMap type considerably.
This type was allocated on the stack in some getTable() calls which
could cause stack overflow for NdbApi users.

This patch dynamically allocates the hashmap from the heap to avoid
this problem.

Dynamic allocation is already used on this path (Vector<>), and this
path should only be taken once per table version per 
ndb_cluster_connection.