WL#7592 step 14. GTIDs: Generate Gtid_log_event and Previous_gtids_log_event. Fix Multi-threaded slave.
Multi-threaded slave has its own logic to parse transaction
boundaries.  Before WL#7592, there was a special case that it handles
in a way that is not compatible with WL#7592.

Before WL#7592, DML is wrapped in BEGIN...COMMIT and DDL normally
consists of a single Query_log_event.  However, there is one
exception: if CREATE TABLE ... SELECT is logged in statement format,
and the SELECT uses a user variable, or calls RAND() or uses
INSERT_ID/LAST_INSERT_ID, then CREATE ... SELECT will result in one or
more User_var_log_event/Rand_log_event/Intvar_log_event, followed by a
Query_log_event.

Before WL#7592, when GTID_MODE=OFF, Multi-threaded slave stored
logical timestamps in the Query_log_event at the beginning of the
transaction.  In the case of CREATE ... SELECT, where the transaction
started with another event, a specialized logic was used.  The master
wrote the Query_log_event with a logical timestamp of 0, since it was
not the first event of the transaction.  The slave then used a
dedicated worker thread (called worker 0) to process any transaction
having timestamp 0, and ensured that all such transactions were
executed in isolation.

The logic of multi-threaded slave, before WL#7592, required that a
User_var_log_event that appeared outside of BEGIN...COMMIT was treated
as a transaction of its own.  This was just a technicality in the code
and did not cause a problem, since the User_var_event and the
following Query_log_event would both be assigned to worker 0.

After WL#7592, there is an Anonymous_gtid_log_event before any CREATE
... SELECT statement, and this event carries the correct logical
timestamps.

Problem:

After WL#7592, the slave may process a sequence of events like:

  Anonymous_gtid_log_event
  User_var_log_event
  Query_log_event

The Anonymous_gtid_log_event has nonzero logical timestamps and
therefore it may be executed by a worker other than worker 0.  The
User_var_log_event will be executed by the same worker.  However,
after WL#7592 and before this patch, the User_var_log_event will
terminate the transaction.  Then the Query_log_event will be
scheduled.  Since it has logical timestamp 0 it will be scheduled to
worker 0, where the User_var_event has not taken effect.  So the wrong
value will be inserted into the table.

In the code, the problem is located to two places:

- In Log_event::get_slave_worker, logic to terminate the transaction
  was invoked for events that appear outside BEGIN...COMMIT and which
  are not COMMIT or Xid.  This includes User_var_log_event appearing
  outside of BEGIN...COMMIT.

- The function slave_worker_exec_job_group iterates over the events of
  a transaction in a while loop.  There was logic to break out of the
  loop that was invoked for events other than
  Gtid_log_event/Anonymous_gtid_log_event which appeared outside
  BEGIN...COMMIT, for the logical clock scheduler.  This includes
  User_var_event appearing outside BEGIN...COMMIT.

Fix:

- In Log_event::get_slave_worker, add a condition so that the logic is
  only invoked for events other than Query_log_event, if a
  Gtid_log_event or Anonymous_gtid_log_event has been seen.

- In slave_worker_exec_job_group, add a condition so that the logic is
  only invoked if there has not been a Gtid_log_event or
  Anonymous_gtid_log_event.