Bug#19143833: DEBUG ASSERTION ON RECOVERY AFTER INSERTING INVALID UUID IN MYSQL.GTID_EXECUTED
Background:

Following are some observations helpful in understanding the issue:

B1: The table mysql.gtid_executed is referred on recovery.
    If a erroneous value is inserted in this table,
    the server refuses to start.

B2: In general when the mysql server sees an error, it logs
    the erroneous status it in the diagnostic area meaning:
    thd->get_stmt_da()->status() == Diagnostics_area::DA_ERROR

B3: The observation in B2 is not always true. The exception being
    that the logging in diagnostic area happens only once the
    server has started and diagnostic area is active. If there is an
    error while the server is still in the process of recovery
    diagnostic area is not active and hence the error is NOT
    reflected in the diagnostic area.

B4: When there is an error in the gtid code, gtid code asserts
    that the error was logged in diagnostic area.


Problem:

This is a case of negative testing. We insert an erroneous
value in the gtid_executed table on purpose and test that
mysql server, according to B1, indeed refuses to restart.

The issue here results from the assert described in B4
ignoring the exceptional case B3 meaning the assert fires
complaining that though there was an error, the diagnostic
area is oblivious to this error.

Fix:

Extend the assert to accomodate the exceptional case.
The assert handles both cases now (joined with an OR):
1) Assert that error is registered in diagnostic area
2) The concerned thread is SYSTEM_THREAD_COMPRESS_GTID_TABLE
   and diagnostic area is empty.