Bug#20379387 TEST THE INNODB DOUBLEWRITE BUFFER WITH VALGRIND
The test for the InnoDB doublewrite buffer is being skipped on Valgrind,
because it uses DBUG_SUICIDE for killing the server. If the server were
killed externally by the test framework, Valgrind would be useable.

We will use kill_mysqld.inc and start_mysqld.inc for killing and starting
the server. The test will still need have_debug.inc for ensuring that
the doublewrite buffer will be able to recover the intentionally corrupted
data.

We will also rename the test from innodb.innodb-double-write
to innodb.doublewrite.

Reviewed-by: Annamalai Gurusami <annamalai.gurusami@oracle.com>