BUG#20380487: RPL_PERFSCHEMA_NO_MASTER_UUID FAILS SPORADICALLY
The test fails due to poor synchronization. The slave thread may
be slow reading the uuid from the master. It starts, connects and
then reads the UUID. The user thread (the one from the test case)
knows nothing about this, it just waits until the slave thread
starts and connects to the master. Thence, it may already be
checking that the UUID exported on the P_S table is the one
expected, and consequently assert too soon - i.e., before the
slave thread actually retrieves the UUID from the master.

The fix is to make sure that the user thread keep checking until
it gets the expected UUID or bails out after a certain amount of
time. The latter will make the test case fail, thence acts as an
implicit assertion.

This patch also cleans up the test case a bit.