Bug #20357597 M_S_I CRASHING FOR USER WITH BLANK/NON-BLANK PASSWORD ON WINDOWS
Description:
execute mysql_secure_installation by default root password which is
blank and observe it is crashing

Analysis:
The part of code responsible for initializing shared memory was
"ifdefed" by HAVE_SMEM. During a refactoring all references to HAVE_SMEM
were replaced by #if defined (_WIN32) && !defined (EMBEDDED_LIBRARY). It
seems that mysql_secure_installation was overlooked during this
refactoring. Also mysql options were not resetuped after mysql_close.
Shared memory is default for windows and that is why the default value
for it was added.