Bug#20219846 EXPLAIN FOR LIMIT QUERY SHOWS FILESORT BUT EXECUTION IS DONE WITH INDEX ONLY
The explain output for the following query:

  SELECT *
  FROM t1 JOIN t2 ON t1.i1=t2.i1
  WHERE t2.i2 > 3
  ORDER BY t1.i1 LIMIT 20;

says the query will be done by doing file sort but when it is executed
it is done by reading from the "order by" friendly index on t1.i1.

The switch to use the "order by" friendly index on t1.i1 is done in
test_if_skip_sort_order() by calling test_if_cheaper_ordering(). The
same test is done in both the explain case and in the execution of the
query. The choice of whether to switch index in
test_if_cheaper_ordering() takes the filtering effect of the tables
into account when computing the expected number of rows that is
needed to read before having produced enough rows to fulfill the
limit criterium. The cause for doing a different choice is that the
filtering effect on the last table is different in these two
cases. When doing explain of a query we always calculate the condition
filtering effect for all tables. When executing the query, we only
calculate the filtering effect of the last table when it is actually
needed. In most cases it is never used but in the case of "order by
limit" optimization it is used (but was not computed).

The fix for this problem is to always calculate the condition
filtering effect the last table if the query has an order by or a group
by clause and a limit clause.

Changes to existing test cases:

myisam_explain_*_non_select_*:
  The printout of status variable from execution now corresponds to the
  explain of the same query.
single_delete_update:
  Adjusted the limit for some queries in order to produce the same
  query plan.
filesort_pq:
  Re-recorded due to changes in rows and filter estimates.