Bug #19974879 SEGMENTATION FAULT IN MYSQL CLIENT WHEN STARTED WITH --XML
MYSQL Client program terminates with Segmentation fault when started
with --xml option and command is executed on client.
Analysis:
The history was not initlized for --xml, however wirtes to the history
were independend of --xml option.
Fix:
Mysql documentation does not say that for xml option logging to the
history should be turned off. That is why I have turned on history
logging for --xml. The same was done for --html option