Bug#17579998 TEST 'TESTNODERESTART -N GCP -L 1 T1' FAILS DUE TO MYSQLD ABORTED
When a data node fails or being restarted, the other nodes in the
same nodegroup will resend data to subscribers that they "believe"
that the failed node have not already sent.

Normally when a data node (suma block) have sent all data for an epoch
it is responsible for, it will send a SUB_GCP_COMPLETE_REP with a count
to all subscribers.  When the subscribers receives SUB_GCP_COMPLETE_REP
it reply with a SUB_GCP_COMPLETE_ACK.  Then Suma have received the ack
from all subscribers it will distribute it to the other nodes in same
nodegroup so they will know that that data need no resend in case the
node fail.  If a node fail between all subscribers sent ack but before
all other nodes in nodegroup received the ack from failing node, data
for some epoch may be sent twice and reported complete twice.  In this
cases a crash as in bug may occur.

This patch add a list of identifiers for which buckets that are reported
in SUB_GCP_COMPLETE_ACK in addition to the count.  The receiver can now
use this list of identifiers to keep track of which buckets are completed
and detect when an already completed bucket are reported a second time, and
can ignore that signal. 

This patch also introduces the concept of Subscriptions Data Streams to hide
the details of Suma buckets in ndbapi.