Bug #20118980 CONNECT FAILURE WHEN TRY TO CONNECT SHA USER WITH PASSWORD OF LENGTH>512 BYTES
Description:
The part of the bug that mysql fails to connect when password length is
>= 512 bytes was discarded by Kristofer as not a bug.

Tha valid part of this bug is the memory leak:
If we execute the above test without specifying client option
MYSQL_SERVER_PUBLIC_KEY then there is also a memory leak of 168 bytes.

Analyzis:
There was a leak of memory allocated for public_key if the password was
to long.
Fix:
The memory for public_key is freed before return if the key is to long.