BUG#19667566 : PARTITION BY WITH COMMENT LEADS TO MYSQL SERVER REBOOT IN MYSQL CLUSTER
If the alter table query has comments and partition clause, it throws
segfault and crashes the mysqld server.

The server calls the handler's can_switch_engines() function to check if
switching engines is allowed when there is an alter table switch engine
or alter partition query. But NDB has native partitioning, so the handler
function has to read into server data(esp create_info) to make decision. But
the value of create_info is sometimes null since the server assigns value to
a local copy and the handler tries to read from the original copy in lex
which is still null. The bug is fixed by adding a null check to create_info.
Changes : 
  - added null check for create_info in ha_ndbcluster::can_switch_engines()
    and removed few redunant checks
  - added test cases to test the fix