Bug#17579998 INVALID SUB_GCP_COMPLETE_REP DUE TO SUMA RESEND DATA FOR ALREADY COMPLETE EPOCH
Post push fix of regression introduced in previous patch.

When a nodegroup was dropped, the SUB_GCP_COMPLETE_REP signals got
a mismatch betweeen number of buckets to report and list of bucket
identifiers in signal.

This was due to that buckets dropped during the preparation of
signal was not included in list of identifiers since it was no
longer an active bucket.  A temporary bitmap keeping track of
dropped buckets previously active was introduced and used to
create the list of identifiers in signal.