Fixed bug#20384201: WL#411: COVERING INDEX IS NOT USED WHEN SELECTING A GENERATED COLUMN
Covering_keys is generated by covering_keys.intersect(field->part_of_key).
When 'field' is a virtual generated column, because it can not be indexed,
part_of_key is always '0'. This results in covering_keys always being 0 if
there is a virtual generated column needed to be read.

Solution: When update covering_keys and merge_keys, ignore virtual
generated columns.