Bug #20075721 	ASSERTION `LOST_GTIDS->IS_EMPTY()' AT BINLOG.CC:3637 ON PURGE BINLOG CMD
We are hitting the assertion when purging binary logs and the first
left binary log just contains a Previous_gtids_log_event. The problem
is caused by making GOT_PREVIOUS_GTIDS case to execute the assertion.
This change is added after fixing Bug#16741603, but missed to
remove the assertion.

Removing the assertion to fix the problem.