Bug#20293609 ST_POINTFROMGEOHASH: CONDITIONAL JUMP OR MOVE DEPENDS ON UNINITIALISED VALUE(S)
ANALYSIS
--------
When a invalid geohash string is given in one of the functions
*FromGeoHash(), an error is raised which prints the invalid geohash
to the client. When printing the error, the function String::c_ptr()
is called on invalid geohash. This function expect the string to
have allocated at least (length() + 1) bytes.

When the invalid geohash is a string that is not null terminated and
the length of the geohash is equal to or greater than the allocated
length, it will cause the mentioned *FromGeoHash() functions to read
uninitialized values when calling c_ptr().

SOLUTION
--------
Use c_ptr_safe() instead of c_prt() which will ensure that the geohash
string is always null terminated.