WL#8250: GCS Replication: MSR adaptation
MSR + MGR (Valgrind Post fix)
THD objects have to be local and so they can't be class fields.
This is fixed in the gcs_applier and gcs_recovery files.

There is also a fix for the proper deletion of channel structures on
server shutdown. This lead to a fix on the rpl_gcs_restart.test as
server restarts can now lead to relay log cleanups.