BUG#20400132 TRANSACTION BOUNDARY PARSER SHOULD NOT DEPEND ON GTID_MODE
It was implemented a transaction boundary parser to help the IO
thread to keep track of transaction boundaries to only add a GTID to
the Retrieved_Transaction_Set when the full transaction was queued.
So, as it was solving a GTID issue, the transaction boundary parser
used by the slave only when gtid_mode >= GTID_MODE_UPGRADE_STEP_2.

As the WL#7083 is now making the GTID_MODE settable on-line, the
boundary parser should be used by the slave regardless of GTID_MODE.

Fix:
===

It was removed the GTID_MODE checks before using the transaction
boundary parser.