Bug#20391871 WEBSCALESQL: DON'T REQUIRE INTERNAL VALUES, THAT SHOULD CHANGE, TO NOT CHANGE
Merge upstream the following contribution from WebScaleSql:

https://github.com/webscalesql/webscalesql-5.6/commit/049aa3335ef9c8e927d125a64e27e7b0ffe4e303?diff=unified

Don't require internal values, that should change, to not change

Summary:
Feature: Basic Test Fixes

1) A number of PS tests show performance_schema_max_statement_classes
along with the other PS variables.  This value is not important, and
this number will just go up every time a new statement class is added.

The only thing testing for this number accomplishes is the creation
of pointless conflicts with all the diffs which introduce these.

So, this one variable is now excluded from the output of those tests.

2) Many internal values are included in the full output of
'show engine performance_schema status', so this has the same issue.

So, these tests were changed to hide the actual numerical values in
the required test output.