BUG#20416013: FAILURES IN RPL.RPL_GTID_SPLIT_STATEMENTS
Fixed the result file. After WL#8206, the ranges
for the error codes were changed. This result file
that was merged and pushed soon after the worklog
was pushed was not updated. This patch updates it,
so that the new error codes are put in the result
file instead of the old ones.