Bug #19625718	OK PACKET HAS 1 EXTRA BYTE OF INFORMATION WHICH IS NOT NEEDED Bug #20126551	MYSQL_SESSION_TRACK_GET_FIRST RETURNS 1 INSTEAD OF 0 IN C/C 6.1.6
This is follow up patch for bug #19625718 which was closed with a patch which altered OK packet layout introduced by the WL#4797.

This patch also fixes Bug #20126551 which was caused by the changes in OK packet layout introduced by fix for Bug #19625718.

With this change the OK packet layout is in sync with the documentation: http://dev.mysql.com/doc/internals/en/packet-OK_Packet.html

Reviewed-by: Rafal Somla <rafal.somla@oracle.com>
Reviewed-by: Bharathy X Satish <bharathy.x.satish@oracle.com>
RB: 7748