Bug#19785977:  CRASH IN NDBINDEXSCANOPERATION::SCANINDEXIMPL
When a transaction is started from a cluster connection, Table and 
Index schema objects may be passed to the transaction for use. If 
these schema objects have been acquired from some other connection, 
they can be deleted at any point by the deletion or disconnection 
of the owning connection. This can leave a connection with invalid 
schema objects, which cause a core when dereferenced.

Added a check to optionally detect sharing of schema objects between 
connections while passing a schema object to a transaction. If the 
check is enabled, the schema objects of the same names are acquired 
from the connection. These are compared to the schema objects passed 
to the transaction. If they do not match, an error is returned.