BUG#20400518 SQL_SLAVE_SKIP_COUNTER WITH CHANNELS CAN BRING               INCONSISTENCY IN RPL
When setting sql_slave_skip_counter, the value is applied to all
channels. So when starting channels, all started appliers will
go into skip process. And the appliers will decrease the global
sql_slave_skip_counter after skipping a event. The logic is not
correct and it will cause problem if two or more channels are
starting at the same time.

To fix it, the be haviour of skipping event is change to:
- Setting sql_slave_skip_counter doesn't effect any running applier.
- Value of sql_slave_skip_counter only effects the first applier
  starting after 'SET GLOBAL sql_slave_skip_counter.
- The applier copies the value to a local variable and just decrease
  the local variable.
- The value of sql_slave_skip_counter will be reset to 0 after
  the applier copies it. So if users need to skip some events
  on another channel, then they they need to set sql_slave_skip_counter
  again.